Refactor braille_writer to use braille_message class
