+++ lib/message.rb
@@ -1,0 +1,6 @@
+class Message 
+  def initialize(message)
+    @message = message
+  end
+
+end
